requirements.txt
pandas>=2.0
numpy>=1.24
pyarrow>=14.0
scikit-learn>=1.3
xgboost>=2.0
mlflow>=2.10
dagshub>=0.3.30
joblib>=1.3
PyYAML>=6.0
fastapi>=0.110
uvicorn>=0.27
requests>=2.31
pytest>=7.4
ruff>=0.3
